+++ lib/mongoid/finders.rb
@@ -7,8 +7,9 @@ module Mongoid
     extend Origin::Forwardable
 
     select_with :with_default_scope
-    delegate :aggregates, :avg, :each, :each_with_index, :extras, :find_and_modify, :for_js,
-      :includes, :map_reduce, :max, :min, :sum, :update, :update_all, to: :with_default_scope
+    delegate :aggregates, :avg, :each, :each_with_index, :extras, :find_and_modify,
+      :first_or_create, :for_js, :includes, :map_reduce, :max, :min, :sum, :update,
+      :update_all, to: :with_default_scope
 
     # Returns a count of records in the database.
     # If you want to specify conditions use where.
@@ -119,19 +120,6 @@ module Mongoid
       with_default_scope.first
     end
 
-    # Find the first +Document+, or creates a new document
-    # with the conditions that were supplied.
-    #
-    # @example First or create the document.
-    #   Person.first_or_create(:attribute => "value")
-    #
-    # @param [ Hash ] attrs The attributes to check.
-    #
-    # @return [ Document ] A matching or newly created document.
-    def first_or_create(attrs = {}, &block)
-      first_or(:create, attrs, &block)
-    end
-
     # Find the last +Document+ given the conditions.
     #
     # @example Find the last document.
@@ -155,19 +143,6 @@ module Mongoid
     # @return [ Document ] The first or new document.
     def find_or(method, attrs = {}, &block)
       where(attrs).first || send(method, attrs, &block)
-    end
-
-    # Find the first document or create/initialize it.
-    #
-    # @example First or perform an action.
-    #   Person.first_or(:create, :name => "Dev")
-    #
-    # @param [ Symbol ] method The method to invoke.
-    # @param [ Hash ] attrs The attributes to query or set.
-    #
-    # @return [ Document ] The first or new document.
-    def first_or(method, attrs = {}, &block)
-      with_default_scope.first || send(method, attrs, &block)
     end
   end
 end